requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
bcrypt==4.2.1
email-validator==2.2.0
gunicorn==23.0.0
